bo/backend/requirements.txt
# FastAPI and related packages
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.9
alembic>=1.10.3
# sqlite3 is a built-in module in Python, not a pip package

# Utilities
python-dotenv>=1.0.0
httpx>=0.24.0
websockets>=11.0.2

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Database Drivers
pymysql>=1.1.0
